changes
added parse args function
adjustment to if__name__ function
I tries to run the code but the arugment file pass in is not recognized for some reason.
Need to fix that
in the def main function, i chnages the readtxt file from readLines() to just read as well as changed some variable name. Test1 is just me trying to make sure the main method will work. to run it  input " python3 test1.py"