Update contactOrganizer.py
added sort methods for name, address, email, and phoneNum. But it wont commit the change to the db browser for some reason, but when you print it out, the output is correct. Need to fix that somehow. Also format the file a little as well as deleted some random imports. Also need to write some test codes.