Create README.md
The challenge: I can never keep track of all my contacts because I have too many
friends. I'm just too popular for my own good.
● Overview of the project. Explain what you propose to create and why.
For this project, we aim to solve problems for people who are too popular or have a
messy contact list. In order to solve this problem, we take in a list of contacts and parse
through it. Using python regex functionalities, we can find specific information to
organize such as phone number, email, name, and addresses.
● Scope of the project. Describe what your project will and will not do.
Convince me that your project is non-trivial yet feasible within the given
timeframe.
Our project will only organize and sort the contact information. We can sort these
information by the name, email, number, or address depending on what the user wants.
We may also include the function of organizing information by ascending or descending
order. However, we will not be able to change/update any of the information. For
example, we will also not be able to add on or delete the information.